Valery's update for root 3.00

+++ StRoot/St_base/table_header.h
@@ -6,11 +6,11 @@
 typedef struct  table_head_st {
    char name[20];	/* table name */
    char type[20];	/* table type */
-   long maxlen;		/* # rows allocated */
-   long nok;		/* # rows filled */
-   long rbytes;		/* number of bytes per row */
-   long dsl_pointer;	/* swizzled (DS_DATASET_T*) */
-   long data_pointer;	/* swizzled (char*) */
+   int  maxlen;		/* # rows allocated */
+   int  nok;		/* # rows filled */
+   int  rbytes;		/* number of bytes per row */
+   int  dsl_pointer;	/* swizzled (DS_DATASET_T*) */
+   int  data_pointer;	/* swizzled (char*) */
 } TABLE_HEAD_ST;
 
 #endif /*TABLE_HEADER_H*/
